Re-add Mill version marker into universal script
Mill used to invoke the JVM with -DMILL_VERSION=… as part of its
universal script. This was used by millw (the Mill wrapper) to
efficiently detect the version of a system-wide mill installation [1].

However, lately this changed and the universal script did not longer
include the -DMILL_VERSION "marker". This re-introduces the marker so
that millw can detect the version of a system-wide mill installation
again.

1: https://github.com/lefou/millw/blob/e62923f3182861d2d1cc0273828f63e5d2ab6763/millw#L108-L139