Derived new GenIdea module to support meta-builds (#2638)
Moved the implementation from `mill.scalalib` to a new module
`mill.idea`. The reason is simple, we need access to the `runner`
module, but `runner` itself depends on `scalalib` to provide compilation
support. Generating an IDE setup has a much broader scope so a dedicated
module is IMHO the best. This is analog to `mill.bsp`, which also
provides IDE support as a dedicated module.

Deprecated the old `mill.scalalib.GenIdea` module.

Enhanced the lookup for modules and root modules to reflect the new
meta-build capabilities of Mill.

Reworked the XML generator to re-respect the
`GenIdeaModule.intellijModulePath` setting and ensured we combine
multiple source folders under the same source root if appropriate.

Pull request: https://github.com/com-lihaoyi/mill/pull/2638