Allow Mill CLI to select the meta-build frame it operates on
Add a new CLI option `--frame` acception an `Int`.
Default in `0` and means the root project,
`1` is the parent meta-build, if defined, or the built-in bootstrap module,
and so on.

This is a first draft, to get more familiar with the recursive but mutable nature of our meta-build support.
Don't hessitate to point out shortcomings.

* Fixes https://github.com/com-lihaoyi/mill/issues/2658

Review by @lihaoyi