Introduce `mill.api.MillException` to transport the error condition (#2789)
Not all build inconsistencies can't be detected before we start
evaluation and not all can be reported as evaluation errors, hence this
PR add an additional way to exit Mill with a proper error message. You
can throw a `mill.api.MillException`. It will be handled in `MillMain`
as well as `MillClientMain`, our central entry points to Mill. The error
message will be reported to the user and Mill exits with code 1. No
stack trace will be shown.

Pull request: https://github.com/com-lihaoyi/mill/pull/2789