Cycle detection in `moduleDeps` and `compileModuleDeps` (#2790)
The problem: We currently don't detect cycles in transitive `moduleDeps`
and `compileModuleDeps`. Cycles in these result in a
`StackOverflowError`.

To solve this, we first need to detect cycles. This is the newly added
`ModuleUtils.recursive` method. Instead of `moduleDeps` we use the
`moduleDepsChecked` in some places to ensure we checked for cycles
before accessing module dependencies.

Unfortunately, we can't easily report the detected cycle at evaluation
time, as `moduleDeps` needs to be evaluated before we can create a task
graph (via `T.traverse`), hence we need an alternative way to propagate
the exception. The current approach throws an `BuildScriptException`
which itself derives from `mill.api.MillException`.

In `MillMain` as well as `MillClientMain`, our central entry points to
Mill, we catch the `MillException` and report the error to the user.

Pull request: https://github.com/com-lihaoyi/mill/pull/2790