fix: Errors not cleaning up in BSP for Scala 3
The Scala 3 compiler is not calling `startUnit` as the Scala 2 compiler.
I checked what sbt is doing and it sends a build/publishDiagnostics
for all the compiled files at the end of compilation.
So I'm doing the same here.
I removed the `CompileProgress` and I'm calling fileVisited for all
input files instead.
I tried to not touch the `CompileProblemReporter` trait for now since
it's public API and I don't want to deal with the binary compatibility
issues. Moreover, our current `fileVisited` API seems more than enough