Avoid recalculating `PathRef`s for `allSourceFiles`
Calculating `PathRef`s is expensive, so reducing them to the minimum
improves performances.
This PR adds an internal task called `allSourceFilesTask` that
is calculated everytime and returns `Seq[os.Path]` instead of
`Seq[PathRef]`. The invalidation still works since `sources`, which is
the input of `allSourceFilesTask` returns the `PathRef`s of the
sources directories