Handle potential null-stream in buildinfo plugin (#2838)
The current Scala implementation does not load the expected values
from the resource file, when the resource file can not be found on the
classpath. The Java implementation already throws a `RuntimeException`
(of an `IOException`) in this case. The Scala implementation does not.

I think missing a generated resource file at runtime is most likely an
runtime issue and we should also throw such exception in the Scala
implementation to exit early, before a missing value can cause greater
harm later. Hence, we now also throw a `RuntimeException` when the
resource file/stream could not be found in the Scala implementation.

Pull request: https://github.com/com-lihaoyi/mill/pull/2838