Propagate env from ctx in Jvm
the current semantics of runSubprocess methods are that the env
arguments are not the whole env but extra arguments passed added to the
system environment.
To allow passing the client env we need to use ctx.
We can't add overloads with an implicit parameter list
so I'm adding private overloads that recieve a `propagateEnv: Boolean`
parameter to be passed to os-lib.
This API is not what we want and will be revisited in Mill 0.12, that's
why I made the new methods private.