Quick fix for issue #2958
Ensure when flatMap-ing, that we don't append duplicates.
Whether this is correct needs yet to be analyzed, but in general,
we don't enfore unique-ness of task results, and even when we evaluate
disjuct tasks, we can't assume that all results are disjuct, hence
using a strict Agg, which enforces uniqueness on append seems not
correct.

Here's my reasoning: If we have two tasks A and B, and B just forwards
the result of A as it's own result, both tasks return the same result.