Fix issue #2958 - deduplicate (anonymous) tasks in result (#2959)
Anonymous tasks can be defined as `val` or `def`. This is in contrast to
named targets, which always need to be defined as `def`.

Also, anonymous tasks end up in the terminal group of the target that's
using them.

If the same anonymous task is used in multiple targets, it also ends up
multiple time in the task results. When a `Evaluator.Results` is
created, this clashes, due to the use of a `Strict.Agg` for
`Evalutor.Results.evaluated`.

This change deduplicates the results before constructing the
`Evaluator.Results`.

Pull request: https://github.com/com-lihaoyi/mill/pull/2959