Prepare the Scala.js integration for SIP-51.
When using Scala.js 1.15.0+ together with Scala 2, we now add an
explicit dependency to the `scalajs-scalalib` artifact. That
artifact comes with Scala.js 1.15.0+, and is double-versioned with
the Scala version and the Scala.js version.

As SIP-51 gets implemented and future versions of the Scala library
may break forward binary compatibility, the new `scalajs-scalalib`
will be able to bring in new APIs back to existing Scala.js versions.

This commit mimics the changes done to sbt-scalajs in
https://github.com/scala-js/scala-js/pull/4913/files#diff-adc7ad642b8f89aab3d2a81ac25b23e345bbea758b1c8bbea46cc9d1f351834c