Print `done` message from test framework to stdout
We are returning the doneMessage from the test framework
as value of the test task, but we are not printing it.
Moreover, the output is shortcircuited when a task fails,
so it's not passed to the `test` command.
This PR uses `ctx.log.outputStream.println(doneMessage)` to
print the message before returning it so users can clearly see it
in case of both success and failure.
This was discovered because of the way weaver works on Scala.js
Since the `TestRunnerTests` are on `scalalib` and can't use Scala.js
code, I implemented two dummy `sbt.testing.Framework`s which just
return a fixed `done` message.