Give mill the ability to re-write ESModule imports at link time  (#3109)
# Motivation

I'm really enjoying frontend scala, without needing to configure an
entire node / npm environment. The capability to use the JS ecosystem
without a bundler, is the "build primitive", that enables this. Here's a
longer discussion of the motivation

https://github.com/VirtusLab/scala-cli/discussions/1968#discussioncomment-5446977

# Implementation
arman added this to SBT here and published a library that does the heavy
lifting.
https://github.com/armanbilge/scalajs-importmap

I followed this up in scala-cli ... 
https://github.com/VirtusLab/scala-js-cli/pull/47

... and am really enjoying this in the small. My larger projects are in
mill though (thanks :-)!). Hence... this PR... which seeks to integrate
the capability into mill. I wanted to do it in a plugin - but I couldn't
see how as the call to the linker is in a private scope - so I've put it
up for mill itself.

I would expect this as is to pass CI with the new test. Very open to
feedback. If accepted, it would be my first contribution to mill ... I'd
be a little surprised if I got everything right straight out the gate -
a plugin may indeed be preferable if I have not correctly understood the
constraints.

Pull request: https://github.com/com-lihaoyi/mill/pull/3109

---------

Co-authored-by: Tobias Roeser <le.petit.fou@web.de>
Co-authored-by: Lorenzo Gabriele <lorenzolespaul@gmail.com>
Co-authored-by: Romain Gilles <58700468+romain-gilles-ultra@users.noreply.github.com>