[BSP] Support buildTarget/wrappedSources for build scripts
scala-cli supports a bespoke extension to the the build server protocol.
This extension allows the support of Scala scripts. Scala scripts
are Scala code which is then wrapped at build time by some Scala
code to make them valid Scala code.
scala-cli-bsp's buildTarget/wrappedSources allows build tools to
give the following information:
- the scala script file
- the generated scala file generated from the script
- the string that is added on the top of the file
- the string that is added on the bottom of the file
This PR adds Virtuslab's scala-cli-bsp library to the bsp worker module,
and implements the handler for the buildTarget/wrappedSources call.
To make the information available, it changes the return type of
`generateScriptSources` in `MillBuildRootModule` to return, together
with the `PathRef` of the destination folder, the four elements
described above, so it can be used by BSP.