Add coarse grained lock around BSP requests to avoid deadlocks (#3243)
Fixes https://github.com/com-lihaoyi/mill/issues/3100, which for some
reason is happening frequently enough on my laptop to really hamper my
productivity. With a version of Mill built from this PR, the problem
seems to have gone away.

We don't really have a strong concurrency model for what's meant to
happen when multiple Mill commands are being evaluated at once, and
trying to come up with one is a complex task that I don't think is worth
the effort. Bazel works fine with a coarse-grained lock even for huge
codebases and organizations, so this should do for Mill for now.

Pull request: https://github.com/com-lihaoyi/mill/pull/3243