Automatically pump `os.proc` streams when `SystemStreams` are redirected (#3275)
Depends on https://github.com/com-lihaoyi/os-lib/pull/283

This moves the subprocess stream handling logic out of
`Jvm.spawnSubprocess` and makes it apply to all `os.proc` invocations,
greatly reducing the room for error. With this, `Jvm.spawnSubprocess`
becomes a very thin wrapper around `os.proc.spawn`. We also rely
directly on OS-Lib's own pumper threads to pump to our destination,
rather than having them pump into in-memory buffers and then spawning
our own pumper threads to pump from those buffers to the destination

I spent some time looking into how to do the stdout/err handling at the
process level, but couldn't find any reasonable mechanism to do so that
allows us to preserve the ordering of the stdout/stderr. This is the
original motivation to squishing it into one stream via
`ProxyOutputStream`/`ProxyStreamPumper` and is important because
otherwise you find e.g. stack traces out of order with printlns, which
makes debugging very difficult. Might be possible using some
socket/fifo/pipe cleverness, but not as part of this PR

Added an integration test to assert on the subtleties of stdout, stderr,
and their inherited alternatives. This PR is required for the test to
pass