Doc: Default to latest stable version (#3351)
With a stable version shown by default, users won't see newly added but
unreleased features.

User interested in a pre-release (e.g. the latest main snapshot build)
can select that version manually, and are able to test the latest
changes.

With this change in effect, we can request proper initial documentation
for feature additions. This not only improves the experience for early
adaptors but also the understanding of the to-be added feature while in
the review process.

Pull request: https://github.com/com-lihaoyi/mill/pull/3351