Make `TestModule#test` run in `T.dest` folder by default (#3347)
This PR makes it such that `TestModule#test` runs subprocesses inside
`T.dest / "sandbox"` by default, rather than `T.workspace`.

This change can be disabled with a flag `def testSandboxWorkingDir =
false`, but is a good default to encourage:

- Before, because tests run with the `pwd` in the repository root, there
are no guardrails preventing you from doing the "easy" thing and just
reading stuff off of disk rather than having explicit build
dependencies.
- Even though often people put stuff in `resources/` and read from it,
which did the right thing "accidentally" since tests depend on the
resource folder, it was easy for people to put stuff in other
`not-resources/` folders and have the dependency between the test files
and the test task missing.
- This also means things like `--watch` do not properly pick up changes,
and `testCached` does not properly get invalidated.
- It also means that it is easy to accidentally write over shared files,
which becomes problematic when parallelism is turned on (which becomes
the default for `testCached` in
https://github.com/com-lihaoyi/mill/pull/3265/files, and we probably
want to encourage more `testCached` going forward)

- With this change, tests run in empty `T.dest / "sandbox"` folders.
This means that it is very hard to "accidentally" do the wrong thing:
- Although it is still possible to read and write anywhere on disk, it
now becomes something you have to try hard to do, rather than something
that can be done by accident.
- Rather, you should prefer to pass files to tests via `forkEnv` or
`forkArgs`, which is a pattern we already do in Mill's own build (e.g.
[here](https://github.com/com-lihaoyi/mill/blob/33fa9530f165a5e5e1d367a5b51b291c2bb15d1f/build.sc#L684-L689)
and
[here](https://github.com/com-lihaoyi/mill/blob/33fa9530f165a5e5e1d367a5b51b291c2bb15d1f/build.sc#L1189-L1193)).
- This makes an dependencies between tests and their files explicit, so
`--watch` works and `testCached` properly invalidates
- To provide convenience for the common case of use cases where we read
stuff from the resource directories, I added a
`MILL_TEST_RESOURCE_FOLDER` environment variable that contains the
resource folders as a semicolon-separated string. In the common case,
this is a single path, and can be used in your test code to refer to the
resource folder as an alternative to the old practice reading directly
from `os.pwd / "src" / "test" / "resources"` or whatever

There is some design leeway here; we could follow bazel, and create
symlinks for all referenced `PathRef`s inside the test sandbox folder.
Or we could do nothing, and let people pass in the stuff they need
manually via `forkEnv` or `forkArgs`. I think the proposed approach of
providing `MILL_TEST_RESOURCE_FOLDER` by default, and letting people
provide other things on their own, is a reasonable compromise between
these approaches. A full design and implementation of a bazel-style
"Runfiles" system is probably more complexity than I am willing bear at
this moment. This change is nowhere near the sophistication of the bazel
sandbox, but it is a start

This is a binary compatible but semantically breaking change, and should
go into 0.12.0. It will also need extensive updates to the
documentation, which I'll postpone till later since we don't want those
changes to be visible to users until this change has been released