Cleanups for Mill client (#3363)
* `ServerFiles.*` and `OutFiles.*` have been turned into constant fields
where possible

* `client/` code layout has been overhauled: `MillClientMain`/`MillEnv`
have been re-organized into `MillClientMain`, `MillLauncher`,
`MillServerLauncher`, `MillNoServerLauncher`

* `MillClientMain.main0`, now living in `MillServerLauncher.runMain`,
has been broken down: `runMillServer` has been extracted out of it, and
the `Locks`/`Lock`/`Locked` logic has been tweaked with the addition of
`TryLocked implements AutoClosable` to let us fold some logic into the
try-with-resources block

* Removed the code path for running the Mill server in-process: this
cuts down the number of execution modes from 3 to 2, server and
no-server. There'll be a bit of performance hit spawning another JVM for
the no-server mode, but IMO it's worth it for the simplicity and
maintainability of this gnarly code, and is necessary for moving the
Mill `os.pwd` out of the root directory into a sandbox

* Some preparations for moving the Mill `os.pwd` out of the root
directory into a sandbox, e.g. introducing a `MILL_WORKSPACE_ROOT`
environment variable and plumbing it throughout the codebase

This should help improve the code and implementation quality of the Mill
client-server codebase and set the stage for further development